Update VICE to revision 25992

+++ jni/VICEPlugin/vice/vdrive/vdrive-command.c
@@ -783,7 +783,7 @@ static int vdrive_command_chpart(vdrive_t *vdrive, BYTE *name, int length)
 
             if ((ss == 0) && ((len % 40) == 0) && (len >= 120) && (ts != 40)) {
                 te = ts + (len / 40);
-                if (((ts < 40) && (te >= 40)) || (te >= vdrive->num_tracks)) {
+                if (((ts < 40) && (te >= 40)) || (te >= (int)vdrive->num_tracks)) {
                     return CBMDOS_IPE_BAD_PARTN; /* FIXME: is this correct ? */
                 }
 
@@ -1169,4 +1169,3 @@ int vdrive_command_memory_read(vdrive_t *vdrive, const BYTE *buf, WORD addr, uns
     vdrive->mem_length = len;
     return CBMDOS_IPE_MEMORY_READ;
 }
-
